Update script runner with new code
