+++ source/Octopus.Tentacle.Kubernetes.ScriptRunner/Program.cs
@@ -1,9 +1,9 @@
 ﻿using System.CommandLine;
 using Octopus.Tentacle.Diagnostics;
 using Octopus.Tentacle.Kubernetes.ScriptRunner.Commands;
+using Octopus.Tentacle.Kubernetes.ScriptRunner.Diagnostics;
 
-//Add the NLog appender (for use with the SystemLog)
-Log.Appenders.Add(new NLogAppender());
+Log.Appenders.Add(new ConsoleLogAppender());
 
 var executeScriptCommand = new ExecuteScriptCommand();
